+++ src/main/java/org/opensearch/neuralsearch/constants/MappingConstants.java
@@ -1,0 +1,15 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+package org.opensearch.neuralsearch.constants;
+
+/**
+ * Constants related to the index mapping.
+ */
+public class MappingConstants {
+    /**
+     * Name for the field type. In index mapping we use this key to define the field type.
+     */
+    public static final String TYPE = "type";
+}